Make thank-you component a root component, import current html, and setup bundle.
- Add account-benefits component
