+++ src/app/main/main.tpl.html
@@ -3,6 +3,7 @@
 <a href="cart.html">Cart</a>
 <a href="sign-in.html">Sign-In</a>
 <a href="checkout.html">Checkout</a>
+<a href="thank-you.html">Thank You</a>
 <ui-view></ui-view>
 
 <div class="test">
